Agregué los archivos end_points en cada paquete api de las apps, donde establezco los end points de las api. También, creé el archivo helpers en la app 'authusers' el cual ahora contiene los métodos create_user_with_credentials, set_token y obtain_token_for_user.
