+++ api_tareas/authusers/api/end_points.py
@@ -1,0 +1,1 @@
+AUTHUSERS_OBTAIN_TOKEN = '/api/v1/authusers/obtain-token'
