+++ api_tareas/tasks/api/end_points.py
@@ -1,0 +1,5 @@
+TASKS_DETAIL_AND_DESTROY = 'tasks:detail_and_destroy'
+
+TASKS_MARK_AS_DONE = 'tasks:mark_as_done'
+
+API_V1_TASKS = '/api/v1/tasks/'
